backend/insecure/requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
PyJWT==2.10.1
requests==2.31.0
pyyaml==6.0
jsonschema==4.19.0
python-dotenv==1.0.0
